+++ src/org/sword/wechat4j/oauth/OAuthManager.java
@@ -10,14 +10,14 @@ import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.util.EntityUtils;
 import org.apache.log4j.Logger;
 import org.sword.wechat4j.common.Config;
-import org.sword.wechat4j.oauth.protocol.getAccessToken.GetAccessTokenRequest;
-import org.sword.wechat4j.oauth.protocol.getAccessToken.GetAccessTokenResponse;
-import org.sword.wechat4j.oauth.protocol.getUserinfo.GetUserinfoRequest;
-import org.sword.wechat4j.oauth.protocol.getUserinfo.GetUserinfoResponse;
-import org.sword.wechat4j.oauth.protocol.refreshAccessToken.RefreshAccessTokenRequest;
-import org.sword.wechat4j.oauth.protocol.refreshAccessToken.RefreshAccessTokenResponse;
-import org.sword.wechat4j.oauth.protocol.validAccessToken.ValidAccessTokenRequest;
-import org.sword.wechat4j.oauth.protocol.validAccessToken.ValidAccessTokenResponse;
+import org.sword.wechat4j.oauth.protocol.get_access_token.GetAccessTokenRequest;
+import org.sword.wechat4j.oauth.protocol.get_access_token.GetAccessTokenResponse;
+import org.sword.wechat4j.oauth.protocol.get_userinfo.GetUserinfoRequest;
+import org.sword.wechat4j.oauth.protocol.get_userinfo.GetUserinfoResponse;
+import org.sword.wechat4j.oauth.protocol.refresh_access_token.RefreshAccessTokenRequest;
+import org.sword.wechat4j.oauth.protocol.refresh_access_token.RefreshAccessTokenResponse;
+import org.sword.wechat4j.oauth.protocol.valid_access_token.ValidAccessTokenRequest;
+import org.sword.wechat4j.oauth.protocol.valid_access_token.ValidAccessTokenResponse;
 
 import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Field;
