+++ src/org/sword/wechat4j/oauth/protocol/refresh_access_token/RefreshAccessTokenResponse.java
@@ -1,17 +1,16 @@
-package org.sword.wechat4j.oauth.protocol.getAccessToken;
+package org.sword.wechat4j.oauth.protocol.refresh_access_token;
 
 /**
- * 响应：通过code换取网页授权access_token
+ * 响应：刷新access_token（如果需要）
  * Created by xuwen on 2015-12-11.
  */
-public class GetAccessTokenResponse {
+public class RefreshAccessTokenResponse {
 
     private String access_token;
     private int expires_in;
     private String refresh_token;
     private String openid;
     private String scope;
-    private String unionid;
 
     public String getAccess_token() {
         return access_token;
@@ -53,23 +52,14 @@ public class GetAccessTokenResponse {
         this.scope = scope;
     }
 
-    public String getUnionid() {
-        return unionid;
-    }
-
-    public void setUnionid(String unionid) {
-        this.unionid = unionid;
-    }
-
     @Override
     public String toString() {
-        return "GetAccessTokenResponse{" +
+        return "RefreshAccessTokenResponse{" +
                 "access_token='" + access_token + '\'' +
                 ", expires_in=" + expires_in +
                 ", refresh_token='" + refresh_token + '\'' +
                 ", openid='" + openid + '\'' +
                 ", scope='" + scope + '\'' +
-                ", unionid='" + unionid + '\'' +
                 '}';
     }
 }
